treefile: fix unit file names
Observed from dmesg since F33 the unit symlinks created with this treefile have invalid names and
are ignored by systemd.

It seems that it worked nonetheless since the units are started anyway and most have correct
symlinks placed alongside the invalid ones. This should get rid of those warnings in dmesg.

Apparently rpm-ostree takes the specified strings here and places them as-is in the .wants/ dir.